Split ReadWriteBufferFromHTTP.h into .h and .cpp file
